Fixed the definition of time_t on NetBSD

+++ include/kernel/netbsd/time.h
@@ -30,7 +30,7 @@ typedef int clockid_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
-typedef int time_t;
+typedef long long time_t;
 # endif
